requirements-ui-tests.txt
playwright>=1.44.0
pytest-playwright>=0.5.0
requests>=2.31.0
nodeenv>=1.8.0
